+++ app/src/main/java/com/example/moviesearchapp/domain/net/data/Movie.kt
@@ -3,8 +3,8 @@ package com.example.moviesearchapp.domain.net.data
 data class Movie(
     val adult: Boolean,
     val backdrop_path: String?,
-    val genre_ids: List<Int>,
-    val id: Int,
+    val genre_ids: List<Long>,
+    val id: Long,
     val original_language: String,
     val original_title: String,
     val overview: String,
